requirements.txt
Flask==0.10.1
requests==2.18.4
python_daemon==2.1.2
